tdf#128651 SdrObjCustomShape::NbcSetSnapRect needs logic rect
SdrObjCustomShape::NbcSetSnapRect needs a logic rect in its parameter.
A general change to snap rect as input in NbcSetSnapRect is not
possible, if the transformation has to be kept. Therefore callers
need to be adapted. This is done her for SdrUndoAttrObj. The error
was that for a transformed custom shape with text, undo and redo had
generated a wrong size of the shape. A least the following actions
were affected: extrusion on/off, shadow on/off, and changes in line
or fill properties. The unit test uses extrusion on.

Change-Id: I8b6990d3d57d25f15f18e8690fc49dbe5e2a2728
Reviewed-on: https://gerrit.libreoffice.org/82367
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>