tdf#128717 Sync duplex value from "Properties" dialog
The value for duplex can be set in two places, the main
print dialog and the "Properties" dialog.

Therefore, update the value for duplex combobox in the main
print dialog with the value set in the PrinterController,
to take into account potential changes carried out in the
"Properties" dialog.

(Other options, like orientation already worked fine before,
as did synchronisation the other way around.)

Change-Id: Ie68f1cfbba6def82c85745e1c361e28a1513e647
Reviewed-on: https://gerrit.libreoffice.org/82528
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>