Fix useless assert(true) (which would never fire)
763c9da2b9ccb39e1cb0d6aed96e4bcf57739535 "loplugin:unnecessaryvirtual" had added
the assert (which was presumably meant to be an assert(false) instead, but had
somewhat oddly been added as part of that commit for no apparent reason), after
9c7097638b74be3ef08fcdf9591fbbdcd4b4693d "made ListBox handle more than 64k
elements, fdo#61520 related" had added the check+return.

But check+assert+return doesn't make sense.  Either the checked condition cannot
happen (assert) or it can (return), but not both.

A `make check screenshot` didn't cause the (now active) assert to fire in my
local build, so lets assume that the condition cannot happen and the assert is
legitimate (and thus drop the return and simplify the code accordingly).

An alternative fix would be to show that the condition can happen, and to thus
drop the assert.

Change-Id: Id0059790d4f34c6645eadda9bca3184483e46bdf
Reviewed-on: https://gerrit.libreoffice.org/82642
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Michael Stahl <michael.stahl@cib.de>