NFC sw: cleanup and make RES_POOLTABSTYLE_DEFAULT consistent
It was introduced in 2016, and shortly thereafter many other
styles were added using the full word TABLE. So make this one
consistent since the eye just glances over the difference.
In fact, in one place a comment even referred it it as
_TABLESTYLE_DEFAULT.
At first I was going to limit the change to just the one
term, but the number of reference for everything was fairly
small, so I just made everything consistent.

Change-Id: I57cc0d081c4bbd5bcea872259065af483c94a334
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/82997
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>