tdf#128739 sw: don't delete bookmarks prematurely in SwUndoDelete
::RedoImpl() and SwUndoInserts::UndoImpl().

The bookmarks will be deleted again by DelContentIndex(), which also
adds them to m_pHistory, but RemoveIdxFromRange() deletes them without
adding them to m_pHistory so they won't be restored.

There is some funny commit c50f2e0a69204b8760c2e06313a18b6194f2d109
which is a mystery: if the rPam has to be saved across the
RemoveIdxFromRange() call, then either the rPam that's passed to
RemoveIdxFromRange() is wrong, or the rPam that is restored after the
call is wrong... just revert it and set the same values to rPam.

Change-Id: I9e93ef75bcd2931594aeae07e761c48752d31d9b
Reviewed-on: https://gerrit.libreoffice.org/83084
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>