sw: insert image: set anchor to at-char by default
This changes the default set in commit
c501b2879983184b6e7ee42b04c38b9c9ba9e9ae (sw: insert image: set anchor
to as-char by default, 2019-07-04), to have a better compromise, taking
both Word defaults compatibility and usability into account.

The problem is that users are used to just inserting an image and being
able to drag it to its final location, which is broken with as-char
anchoring.

So default to at-char anchoring, this is still something that is fully
interoperable to Word (unlike the old to-para anchoring), but allows the
easier image move again.

Change-Id: Ibc61ae167fc9e5cc31b04c83e854556309e27fd4
Reviewed-on: https://gerrit.libreoffice.org/83089
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins