SpellingPopup: lok: Hide unusable items for online
MN_EXPLANATION_LINK is some link, but it's exectued internally,
which means the the link would be executed on the serves side and
not on the client's browser.
m_nCorrectMenuId is a sub menu which allows to specify new autocorrect
rules. Since m_nAddId is also disabled, I expect that extending
the dictionaries / autocorrect rule might be a problem for online.

Change-Id: Id6e8ed078f90a4f25a335a3ab5ec0e3056db648d
Reviewed-on: https://gerrit.libreoffice.org/83596
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>
Tested-by: Tamás Zolnai <tamas.zolnai@collabora.com>