sw: RTF export: fix invalid RTF for shapes and text frames again
It turns out that commit ed9676d4c358112733f0655502f0f2cee7dd7443 wasn't
quite right: writing out the m_aRunText doesn't necessarily put it in
the right place, it just happened to work on that bugdoc but in other
cases the shape may be written before the paragraph properties of the
paragraph it's in...

So instead change the text frame case to save and restore the run
related buffers and members, so that the text frame itself can do
whatever it wants, and the result is appended to m_aRunText.

The members saved are a superset of those saved in
WriteHeaderFooter_Impl(), writeTextFrame(), TextFootnote_Impl().

Also another thing was missing is that OLE objects would trigger the
assert because they are also buffered into m_aRunText, e.g.
ooo46760-1.odt.

Change-Id: Icc6988f8ed19724c593760df46e0254792cd1735
Reviewed-on: https://gerrit.libreoffice.org/83105
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>