Resolves: tdf#108452 invisible all-black drawing cursors on black bg
commit 458cfa88963658da87936dce6bb05d8299d35190
Date:   Fri Feb 8 10:00:12 2019 +0200

    make cursor images part of the theme

+

commit 4bccb6edd840b61d5c5508d2d200251a4d5a4fca
Date:   Mon Mar 4 15:58:18 2019 +0200

    tdf#123796 Wrong cursor when inserting shape

effectively resulted in the gtk version using the same curros as the mac
version. Some of these icons are all black and invisible on a black background,
which affects both gtk and osx.

Taking a copy of LibreOffice 6-1 where the classic xbms were in use, dump
all the xbm cursors and masks as loaded via cairo to png and where the mac
cursors were all black take the xbm derived versions instead (if they were of
the right size)

Change-Id: I8a7a150caff415fa25180c1f2ed4cb8487fcd70c
Reviewed-on: https://gerrit.libreoffice.org/83112
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>