sw: revert change in expanding hints in SwXText::insertTextContent()
The SwXText implementation for inserting text works like this:

* XTextPortionAppend methods appendTextPortion()/insertTextPortion() get
  the text properties passed as a parameter, and they should apply only
  those properties to the inserted text, not expand any existing
  formatting hints at the insert position.

* XSimpleText method insertString() does expand existing formatting
  at the insert position, just like editing in the UI does

For inserting XTextContent:

* XTextContentAppend methods
  appendTextContent()/insertTextContentWithProperties()
  with properties parameter, similar to XTextPortionAppend

* XTextContent::insertTextContent(), without properties

So arguably, by analogy to inserting text, the methods that take
properties should not expand hints, and the insertTextContent() should
follow the insertString and expand hints.

Commit 8a50af25373154a0f1d0bc5e70cb9309aaad84af is an important bugfix
for writerfilter import, but the problem is, it added the
DontExpandFormat() call to insertTextContent(), whereas the regression
it was fixing (from commit 86afb613817e4ef1895f2a2e90d570403c8336cb) was
that the call was removed from insertTextContentWithProperties().

So restore the state before 86afb613817e4ef1895f2a2e90d570403c8336cb.

Turns out that SwUiWriterTest2::testTdf126206 was checking how a
bookmark-start is formatted instead of how the text is formatted.

Change-Id: If524409f88a1a36aa062b6e132996d3f9c1bb571
Reviewed-on: https://gerrit.libreoffice.org/83223
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>