xmloff,sw: fix handling of invalid attribute with multiple ':'
ooo53770-1.odt contains an invalid attribute
style:style:use-optimal-row-height, which was round-tripped as an
unknown attribute, triggering the assert in CheckValidName().

This reveals a confusing mess of SvXMLNamespaceMap::GetKeyByAttrName
functions.

The bCache flag for one of them was accidentally inverted in commit
24308b57c085e6fd001920f065ecdcf7b2a257b4 "loplugin:constantparam in xmloff/"

There are basically 2 use cases for this:
* XML element and attribute names, which are XML QName and contain at
  most 1 ':'
* XML attribute values, which are namespace-prefixed and may contain
  any number of ':' in the "local" part of the value

Because caching is explicitly disabled for Calc formulas, just resolve
the inconsistent bCache by always caching element and attribute names,
and never caching attribute values, so we need just 1 flag.

Change-Id: I363ea2229d0bf5c7199d61b0fee0d9f168911bfa
Reviewed-on: https://gerrit.libreoffice.org/83619
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>