tdf#128629 vcl DrawTransformedBitmapEx: do scaling for flipped bitmaps
Regression from commit b377f8e0979b9c1d186081482da82d60f6584f62 (vcl:
avoid downscale && upscale in DrawTransformedBitmapEx(), 2019-10-08),
the original problem to be solved was that in case you downscale a
bitmap and upscale it later, then you get blurry result, so we try to
avoid touching the pixels and just scale during rendering of the bitmap.

However, here the problem is that scaling is also (mis)used for flip
purposes, so go back to the original behavior for negative scaling.

This keeps the original problem fixed and solves the loss of flip as
well.

Change-Id: Ic9a6eb49d55f2fb8ccf18d982e574398f010cabd
Reviewed-on: https://gerrit.libreoffice.org/83711
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>