partial revert tdf#127316 for rtfexport: restore ++nProp
Because the nProp is multiplied by 100, adding +1 can have
special meaning, as also noted in editeng/impedit4.cxx.

This is an interesting hack, where _nProp_ is used
to determine if _nEsc_ should be automatic.

It was dumb of me to even consider getting rid of it just
because I didn't understand it.
So revert that part of the change.

Change-Id: Ibae5a84c369474f00702614ac5eecdc533aea914
Reviewed-on: https://gerrit.libreoffice.org/83846
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Justin Luth <justin_luth@sil.org>