do not claim in titlebar test that OpenGL is enabled if it's not
Skia takes precedence, if enabled. And I'm not going to add Skia text
to the titlebar, it'd need to include whether it's raster or vulkan,
which would be longer, and I don't see why this information needs
to be in the titlebar in the first place (or if it needs, then
it should be consistent and include also gtk3, qt5, etc.).

Change-Id: I9b95c1a986a38293ee65ee49d1a1970c75b3d202