WIN enable osl_File.cxx part of CppUnitTest_sal_osl
This just disables open::open_004 test on Windows, as this returns
E_NOENT on Windows, but I think the assert is correct. There is
also test-getsystempathfromfileurl.cxx, which seems to overlap?

The test now uses osl::FileBase::getTempDirURL to find the system
temporary directory, insted of hardcoded defaults.

This also includes reverts the following commits:

* "remove unused Exist from oslCheckMode enum"
  commit 89c5abe788ca47a4f321c4b38148544775b68dcd.
* "CppunitTest_sal_osl_file fails on Windows"
  commit 59cf46de31c00c943a287fa77505ed52e47a1ffb

Change-Id: Ib7fc4c13f8159a240d4e20c26f61d84204e4261b
Reviewed-on: https://gerrit.libreoffice.org/83901
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>