Resolves: tdf#129082 insert sheet from file fails
since...

tdf#127958 crash adding 2 or more files into writer master document
commit d988c4e5aefd01b0d091f1e31fc0fd799a9355d3
Date:   Mon Oct 7 09:40:41 2019 +0200

used the SfxItemSet copy ctor to make a copy of the itemset, but the orig
itemset was a SfxAllItemSet not a SfxItemSet.

an SfxAllItemSet allows all items to be added but a plain SfxItemSet only
allows items in certain ranges to be inserted. Upshot is that later on the
filename property cannot be set in the itemset, so file can't be opened.

since...

tdf#128898 import text from file with text-decode failed
commit 3e89f00b51127ab3fd61c3c3f952bb42df243e12
Date:   Wed Nov 27 11:47:57 2019 +0200

    tdf#128898 import text from file with text-decode failed

these are std::shared_ptr so we can return to passing in the original
SfxAllItemSet again to fix the problem

Change-Id: I63a4230a96f9dae7726d854297ebb2b51df4868d
Reviewed-on: https://gerrit.libreoffice.org/84024
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>