Make OpenCLMinimumDataSize configuration property non-nillable
It had curiously been added as nillable (by default) alongside other properties
that were added as non-nillable with 455fdd64617c2a8d7734f6f639a128f5591bfc20
"More work on the new OpenCL options", but with a non-nil default value, and no
code sets a nil value, and ScModelObj::getFormulaCellNumberLimit, which reads
its value and expects it to be non-nil, would cause UB if it was nil.

So make it non-nillable and thereby simplify the code handling it.  (Similar to
how b7353ff83790ca40a631ffc8191c393e8e50a078 "Fixed cppheader.xsl nillable
treatment" retrofitted lots of existing properties to be non-nillable in the
past.)

Change-Id: I60ac95025f363e346f3102ef80f38d00862e6e9a
Reviewed-on: https://gerrit.libreoffice.org/84248
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>