tdf#129071 Qt5 handle windows with parents as dialogs
This is the main fix for this bug. Qt Xcb is a little picky here.
It won't tell the window manager about the transient state, using
WM_TRANSIENT_FOR, for all Qt::WindowTypes. The QPA internal list
(see isTransient function) doesn't include the Qt::Window, which
is qt5 VCL's primary used window type.

This has two consequences:
1. LO dialogs show up in the plasma task manager as seperate
   windows.
2. LO has to handle the transient state itself, instead of relying
   on the window manager. This results in flickering, because LO
   can just push a transient window to the top again, after the
   parent window was activated and therefore drawn in front.

So this just declares all windows with parents to be dialogs.
Almost everything else should be a top-level window anyway. If
not, there is SalFrameStyleFlags::DIALOG to create a parent-less
dialog window.

Change-Id: I89a6d46fd09e4f1d1d2904e152a26733eb266079
Reviewed-on: https://gerrit.libreoffice.org/84298
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>