android: Allow zooming for Calc as well
The comment mentioning "Calc has a fixed zoom at 1x
and doesn't allow zooming for now" was there since
support for the Calc UI was added in commit
1a7db80e089be9ac6aeeec90978a8278046122ca  ("Calc UI
on Android Viewer") in 2017. That seems no longer to be
the case however, probably since commit
62c3ae392efe8b91eef5d28fe75f08192e64b44a
("sc lok: Implement hi-dpi and zoom for spreadsheets.").

Zooming in and out was working just fine when I tested this
with Android Viewer using a spreadsheet containing some text
and numbers, as well as a chart and an image.

Therefore, allow zooming for Calc as well.

Don't allow to set min zoom below the page width for Calc as
somewhat suggested by the existing comment ("applies to all
types of document; in the future spreadsheets may be
singled out"). I don't see a particular reason for this and
using 0f for minZoom would break double-tap zoom.

Change-Id: Ib08439edbbac59a00005213b533df5792da965ee
Reviewed-on: https://gerrit.libreoffice.org/84421
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>