Related: tdf#115719 DOCX import: fix increased spacing vs left-aligned objects
Commit 7e58ed4bdca81715bd6bbcfa9e29b0b90c8c2237 (tdf#115719 DOCX import:
increase paragraph spacing for anchored objects, 2018-02-14) added an
import-time tweak for a problem that has been confirmed to be a Word
layout bug in the meantime (and the tweak makes Writer behave the same
way if the document has been created by an affected Word version for
layout compatiblity).

Later, commit 5bcc89930f545b714898ff346d752c4841f2dba2 (Related:
tdf#124600 DOCX import: ignore left wrap on left-aligned shapes,
2018-02-14) fixed left spacing of anchored objects aligned to the left,
to be in sync with what the DOC import does.

This broke the previous fix in case the shapes are left-aligned.

Fix the problem by tracking what is the in-file-format and logical left
margin, so the final doc model has the value necessary for correct
horizontal positioning and the importer has the value that's necessary
for correct vertical positioning.

Change-Id: I8f16cbe7bad40e243111c902bdc1ab0e8141d6b9
Reviewed-on: https://gerrit.libreoffice.org/84654
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins