ofz#18563 sw: remove SwXFrame::m_pCopySource to fix ~SwIndexReg assert
The problem is that a text frame is inserted with a selection that
partially selects a fieldmark, so the DeleteAndJoin() is split across
multiple calls with a temporary cursor, and the correction of the passed
in rPam that is ultimately SwXFrame::m_pCopySource is not done, so it
remains on a deleted node.

Replace it with a SwUnoCursor, which is automatically corrected.

It turns out that this m_pCopySource member was only set via a call to
SetSelection() from one place in SwXText::convertToTextFrame(),
which ends up immediately calling SwXFrame::attach() anyway, which then
uses m_pCopySource and resets it.  This was added in
deba85c5b73f36affe672567ed4a45938953f312 but it would be far simpler to
just pass a local variable to SwXFrame::attachToRange().

Change-Id: I85ed128db63c13f81f215d8f50036d9d2aa6c519
Reviewed-on: https://gerrit.libreoffice.org/84661
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>