sidebar: no scrollbar in LOK
Scrolling in LOK is done in the client. We render
the sidebar as long as the original height is
(the minimum that will fit all panels) and scroll
in the client. For that, we keep the scrollbar
disabled when layouting the Deck.

Reviewed-on: https://gerrit.libreoffice.org/84720
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>
(cherry picked from commit 3bcd586359b2a8ce6dad3d72c66f26f0f43a6958)

Change-Id: If0d813b04af679768f80ee88fc976bce9cde46d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85785
Tested-by: Jenkins
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>