tdf#129173 tdf#129175 Fix number format of data labels
Show the real value (cell value) of datapoints instead
of the recalculated value in case of Percent Stacked Area
chart.

Also fix: tdf#95425
Do not reset number format of data labels when open dialog
of 'Format data series' and close dialog.

Note: Inherits the data series/point label format from the
cell format and not the axis format, if we set the 'link to source data'
option to true.

Change-Id: I04e9968034a67c6bc6b92941df61d945b4292d79
Reviewed-on: https://gerrit.libreoffice.org/84819
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>