Revert "Enable/Disable of children shouldn't be necessary"
This currently breaks the child sensitivity handling of frames
used in the PDF export, when toggling the form export check box.
I also tried to add a set_sensitive to SalInstanceContainer,
but that didn't work.

I found this while debugging tdf#127493, but I'm don't know, this
is the correct fix wrt welding.

This reverts commit 67160934d6411fc1aa8d1d2fe93da376cadab9bd.

Change-Id: Ib468c82124dc3d1030d7b08b2e139f49f3c00630
Reviewed-on: https://gerrit.libreoffice.org/84976
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>