Prevent BroadcastRecalcOnRefMoveHandler copies
...which would contain copies of AutoCalcSwitch and ScBulkBroadcast that both do
things in their dtors.  std::sort (into which BroadcastRecalcOnRefMoveHandler is
passed as function object argument) returns its function object argument, which
caused the BroadcastRecalcOnRefMoveHandler copies.

Better split the RAII part that holds AutoCalcSwitch and ScBulkBroadcast and
does things in their dtors into a BroadcastRecalcOnRefMoveGuard class of its
own.  (And while at it, explicitly delete the copy/move functions of
AutoCalcSwitch and ScBulkBroadcast to avoid such issues in the future.)

(The RAII part of BroadcastRecalcOnRefMoveHandler had been added with
6c81f2984f92b66dc39e6ecee642b5984941da18 "sc-perf: tdf#87101 add bulk scope for
BroadcastRecalcOnRefMove() calls".  The issue was found with Clang 10 trunk
-Wdeprecated-copy-dtor.)

Change-Id: I45d6b81e7c8b34aed57b6f66c5e1e966a43a565d
Reviewed-on: https://gerrit.libreoffice.org/85063
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>