android: Fix linking of liblo-native-code.so on x86_64
This is similar to commit e71c95737932438c9e1ad858c2da58869267da13
("commit android: Fix linking of liblo-native-code.so on aarch64.").

Build previously failed with

    Linking obj/local/x86_64/liblo-native-code.so
    .../x86_64-linux-android/bin/ld.gold: error: cannot find -landroid_support
    clang++: error: linker command failed with exit code 1 (use -v to see invocation)
    make[2]: *** [../Bootstrap/Makefile.shared:62: obj/local/x86_64/liblo-native-code.so] Error 1
    make[1]: *** [.../android/CustomTarget_lo_android.mk:18: .../workdir/CustomTarget/android/source/done] Error 2
    make: *** [Makefile:282: build] Error 2

I quickly checked that the app actually runs on an x86_64 AVD
and is able to open and display documents after successfully
building with this commit in place.

Change-Id: I92b3759af6a86a6717c287d035b6a1add3a9af7a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85204
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>